Move push notification setup to _Layout
